requirements.txt
clickhouse-connect>=0.6.0
celery>=5.3.0
celery-singleton>=0.3.0
redis>=5.0.0
loguru>=0.7.0
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
cdlib>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0